Initial restructure - tests broken atm
