+++ package.json
@@ -2,10 +2,10 @@
   "name": "Async-Voter",
   "version": "1.0.0",
   "description": "For Voting on Stories and Tickets remotely and asynchronously",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js",
+    "start": "node server.js",
+    "dev": "nodemon server.js",
     "test": "nyc mocha test.js",
     "test:watch": "mocha -w test.js",
     "test:debug": "mocha --inspect --debug-brk test.js"
